Add files via upload
data from the wind explorer of the University of Chile, which shows information about the wind resource based on the use of numerical modeling. The developed modeling provides data simulated with a mesoscale atmospheric model, independently of local meteorological stations. The model used is the WRF (Weather Research and Forecasting) version 3.2, which has been developed by NCAR (National Center for Atmospheric Research) in the United States and is widely used in the field of wind resource assessment worldwide. The model was applied with a spatial resolution of 1 kilometer and has 12 vertical levels between 0 and 200 meters high. Data are available for the entire year 2010.

Source: http://walker.dgf.uchile.cl/Explorador/Eolico2/